게시물 수정, 삭제 권한 설정 기능 구현

+++ src/main/webapp/jsp/article/detail.jsp
@@ -6,6 +6,7 @@
 
 <%
 	Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 
 <html>
@@ -19,10 +20,17 @@
 	<div>번호 : <%= (int) articleRow.get("id") %></div>
 	<div>날짜 : <%= (LocalDateTime) articleRow.get("regDate") %></div>
 	<div>제목 : <%= (String) articleRow.get("title") %></div>
+	<div>작성자 : <%= (String) articleRow.get("loginId") %></div>
 	<div>내용 : <%= (String) articleRow.get("body") %></div>
 	
 	<button><a href="list" style='color : inherit; text-decoration : none;'>목록</a></button>
 	<button><a href="modify?id=<%= articleRow.get("id") %>" style='color : inherit; text-decoration : none;'>수정</a></button>
+	<%
+	if (loginedMemberId == (int) articleRow.get("id")) {
+	%>
 	<button><a href="doDelete?id=<%= articleRow.get("id") %>" style='color : inherit; text-decoration : none;' onclick="if(confirm('삭제하시겠습니까?') == false) return false;">삭제</a></button>
+	<%
+	}
+	%>
 </body>
 </html>
